+++ include/spatial_search_result.h
